refactor(frontend): Update Location component to use Mapbox API for reverse geocoding
